Update static website generation namespaces and process

+++ build.sh
@@ -1,10 +1,16 @@
 #!/bin/bash
 
-rm -fr resources/public/js/*
-clj -A:js
-cp src/cljs/choices/custom.cljs docs/exemples/clojure.txt
-cp resources/public/js/choices.js docs/js/
-echo "static js generated"
+if [[ $1 == "all" ]]; then
+   rm -fr resources/public/js/*
+   clj -A:js
+   cp src/cljs/choices/custom.cljs docs/exemples/clojure.txt
+   cp resources/public/js/choices.js docs/js/
+   echo "static js generated"
+fi
 
-clj -m choices.generateall
-echo "static website generated"
+# Build json files
+clj -m choices.json
+
+# Build index.html
+clj -m choices.index
+
